UPDATED FIGURES FOR PAPER. Added figures to cover the new parts of the paper, with more supplementary figures and fewer principal figures.
ADDED (a), (b), (c), LABELS TO TICK_DIRS.
Added a command to the tick_dirs.py function which allows for lettered annotation to be added to the figures automatically.

ADDED ARROWS TO Dd18O FIGURES.
Added arrows between the maximal Dd18O and the 0 Dd18O to show how this can be a proxy for NDPW export.